\name{best_splits_interface}
\alias{best_splits_interface}
\title{best splits interface}
\description{Compute a data.frame with one row for each segment, and columns
splits and depth, number/depth of candidate splits that need to be
computed after splitting that segment.}
\usage{best_splits_interface(n_data, 
    min_segment_length)}
\arguments{
  \item{n_data}{
}
  \item{min_segment_length}{
}
}



\author{Toby Dylan Hocking}





