\name{cum_median}
\alias{cum_median}
\title{cum median}
\description{Efficient log-linear cumulative median.}
\usage{cum_median(data.vec, 
    weight.vec = rep(1, 
        length(data.vec)))}
\arguments{
  \item{data.vec}{Numeric vector of data.}
  \item{weight.vec}{Numeric vector of weights.}
}



\author{Toby Dylan Hocking}





