\name{depth_first_interface}
\alias{depth_first_interface}
\title{depth first interface}
\description{Use depth first search to compute a data.frame
with one row for each segment, and columns
splits and depth, number/depth of candidate
splits that need to be
computed after splitting that segment.}
\usage{depth_first_interface(n_data, 
    min_segment_length)}
\arguments{
  \item{n_data}{n_data }
  \item{min_segment_length}{min_segment_length }
}



\author{Toby Dylan Hocking}





