\name{get_best_heuristic_equal}
\alias{get_best_heuristic_equal}
\title{get best heuristic equal}
\description{Compute a fast approximate best case based on equal size splits.}
\usage{get_best_heuristic_equal(N.data, 
    min.segment.length)}
\arguments{
  \item{N.data}{
}
  \item{min.segment.length}{
}
}



\author{Toby Dylan Hocking}





