\name{get_complexity_best_heuristic_equal_breadth_full}
\alias{get_complexity_best_heuristic_equal_breadth_full}
\title{get complexity best heuristic equal breadth full}
\description{Compute a fast approximate best case based on equal size splits.}
\usage{get_complexity_best_heuristic_equal_breadth_full(N.data, 
    min.segment.length)}
\arguments{
  \item{N.data}{N.data }
  \item{min.segment.length}{min.segment.length }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





