\name{get_complexity_best_optimal_splits}
\alias{get_complexity_best_optimal_splits}
\title{get complexity best optimal splits}
\description{Convert output of \code{\link{get_complexity_best_optimal_tree}} to counts of
candidate splits that need to be considered at each iteration.}
\usage{get_complexity_best_optimal_splits(node.dt, 
    min.segment.length)}
\arguments{
  \item{node.dt}{node.dt }
  \item{min.segment.length}{min.segment.length }
}

\value{Data table with one row for each segment.}

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





