\name{get_complexity_worst}
\alias{get_complexity_worst}
\title{get complexity worst}
\description{Get full sequence of splits which results in worst case time
complexity.}
\usage{get_complexity_worst(N.data, 
    min.segment.length)}
\arguments{
  \item{N.data}{N.data }
  \item{min.segment.length}{min.segment.length }
}



\author{Toby Dylan Hocking}





