\name{get_distribution_info}
\alias{get_distribution_info}
\title{get distribution info}
\description{Compute a data.frame with one row for each distribution
implemented in the C++ code, and columns distribution.str,
parameters, description.}
\usage{get_distribution_info()}



\author{Toby Dylan Hocking}





