\name{plot.binseg_normal}
\alias{plot.binseg_normal}
\title{plot binseg normal}
\description{Plot loss values from binary segmentation.}
\usage{\method{plot}{binseg_normal}(x, 
    ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg_normal}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking}





