\name{plot.binseg_normal_cv}
\alias{plot.binseg_normal_cv}
\title{plot binseg normal cv}
\description{Plot loss values from binary segmentation.}
\usage{\method{plot}{binseg_normal_cv}(x, 
    ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg_normal_cv}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking}





