\name{plot.binsegRcpp}
\alias{plot.binsegRcpp}
\title{plot binsegRcpp}
\description{Plot loss values from binary segmentation.}
\usage{\method{plot}{binsegRcpp}(x, ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





