\name{plot.complexity}
\alias{plot.complexity}
\title{plot complexity}
\description{Plot comparing empirical number of splits to best/worst case.}
\usage{\method{plot}{complexity}(x, ...)}
\arguments{
  \item{x}{data.table from \code{\link{get_complexity}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





