\name{plot.splits}
\alias{plot.splits}
\title{plot splits}
\description{Plot comparing empirical number of splits to best/worst case.}
\usage{\method{plot}{splits}(x, ...)}
\arguments{
  \item{x}{data.table from \code{\link{get_splits}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking}





