\name{print.binseg_normal}
\alias{print.binseg_normal}
\title{print binseg normal}
\description{Print method for \code{\link{binseg_normal}}.}
\usage{\method{print}{binseg_normal}(x, 
    ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg_normal}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking}





