\name{print.binseg_normal_cv}
\alias{print.binseg_normal_cv}
\title{print binseg normal cv}
\description{Print method for \code{\link{binseg_normal_cv}}.}
\usage{\method{print}{binseg_normal_cv}(x, 
    ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg_normal_cv}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





