\name{print.binsegRcpp}
\alias{print.binsegRcpp}
\title{print binsegRcpp}
\description{Print method for binsegRcpp.}
\usage{\method{print}{binsegRcpp}(x, ...)}
\arguments{
  \item{x}{data.table from \code{\link{binseg}}.}
  \item{\dots}{ignored.}
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





