\name{size_to_splits}
\alias{size_to_splits}
\title{size to splits}
\description{Convert segment \code{size} to number of splits which must be computed
during the optimization.}
\usage{size_to_splits(size, 
    min.segment.length)}
\arguments{
  \item{size}{Segment \code{size}, positive integer.}
  \item{min.segment.length}{Minimum segment length, positive integer.}
}

\value{Number of splits, integer. }

\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





