\name{binseg_interface}
\alias{binseg_interface}
\title{binseg interface}
\description{Low-level interface to binary segmentation algorithm.}
\usage{binseg_interface(data_vec, 
    weight_vec, max_segments, 
    min_segment_length, 
    distribution_str, 
    is_validation_vec, 
    position_vec)}
\arguments{
  \item{data_vec}{
}
  \item{weight_vec}{
}
  \item{max_segments}{
}
  \item{min_segment_length}{
}
  \item{distribution_str}{
}
  \item{is_validation_vec}{
}
  \item{position_vec}{
}
}



\author{Toby Dylan Hocking}





