\name{binseg_interface}
\alias{binseg_interface}
\title{binseg interface}

\usage{binseg_interface(data_vec, 
    weight_vec, kmax, 
    distribution, is_validation_vec, 
    position_vec)}
\arguments{
  \item{data_vec}{
}
  \item{weight_vec}{
}
  \item{kmax}{
}
  \item{distribution}{
}
  \item{is_validation_vec}{
}
  \item{position_vec}{
}
}



\author{Toby Dylan Hocking}





