\name{check_sizes}
\alias{check_sizes}
\title{check sizes}
\description{Checks types and values of size inputs.}
\usage{check_sizes(N.data, min.segment.length, 
    n.segments)}
\arguments{
  \item{N.data}{N.data }
  \item{min.segment.length}{min.segment.length }
  \item{n.segments}{n.segments }
}



\author{Toby Dylan Hocking <toby.hocking@r-project.org> [aut, cre]}





