\name{get_complexity_best_heuristic_equal_depth_full}
\alias{get_complexity_best_heuristic_equal_depth_full}
\title{get complexity best heuristic equal depth full}
\description{Heuristic depth first.}
\usage{get_complexity_best_heuristic_equal_depth_full(N.data, 
    min.segment.length)}
\arguments{
  \item{N.data}{
}
  \item{min.segment.length}{
}
}



\author{Toby Dylan Hocking}





